Improve undelete searches (individual files).  Add new function.
The algorithm for searching for deleted files of specific names can
turn up the same transaction ID multiple times, and also it checks
transaction IDs out of order.  Update the logic to skip TIDs
that are older than the latest successful TID.

Also add a function to produce a timestamp based on the modification
time.  This is used for deleted directories in a future commit.