Add search-directory capability
Now when slider is passed the pass of an existing directory, that
directory's history will be searched for all deleted files and
subdirectories that can be resurrected.  The undelete candidates will
be listed in two groups (directories first) and selectable.

If selected, a file can be undeleted or saved to a new filename.  A
selected directory can be completely restored or saved to a new
directory.

What is still missing is the ability to browse a deleted directory
to see what files are in it (but the individual files won't be
browsable.