Improve history menu for existing file
After extended testing, I found out that not all the provided transaction
IDs are valid.  Each one has to be test for FIFO and that it will open.
On the test for .joe_state, I originally had 53 entries.  After applying
this filter, It dropped to about half of that.

I also find two instances of opening a file descriptor but not closing it,
so this change also stops leaking file descriptors.

Bump to version 2.02