Fix mysterious curses bug on ghost menu where F4: Quit was dimmed
I tried 30 ways to fix this.  I think in the end, the bug was caused by
not assigning "bright" to an initial color like "c_white".  By the time
I figured that out (and it's still not confirmed), I had already rewritten
the menu code.  It's more efficient -- I don't know about more readable
though.  In any case, it works now and probably better, and the F4 option
is now properly brightened.